Create new Crash Report test class and move the previously created UI smoke tests and refactoring work
