+++ app/src/androidTest/java/org/mozilla/fenix/ui/robots/BrowserRobot.kt
@@ -750,6 +750,8 @@ class BrowserRobot {
             val tabCrashedCloseButton = mDevice.findObject(text("Close tab"))
             tabCrashedCloseButton.click()
 
+            mDevice.waitForIdle()
+
             HomeScreenRobot().interact()
             return HomeScreenRobot.Transition()
         }
